use existing template for test
