--- src/mono/wasm/templates/templates/wasi-native/.template.config/template.json
@@ -1,34 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/template",
-  "author": "Microsoft",
-  "classifications": [ "Wasi", "WasiNative" ],
-  "groupIdentity": "Wasi.Native",
-  "precedence": 9000,
-  "identity": "Wasi.Native.9.0",
-  "name": "Wasi Native App",
-  "description": "A project template for creating a .NET app that runs on a WASI runtime",
-  "shortName": "wasinative",
-  "sourceName": "wasi-native.0",
-  "preferNameDirectory": true,
-  "tags": {
-    "language": "C#",
-    "type": "project"
-  },
-  "symbols": {
-    "framework": {
-      "type": "parameter",
-      "description": "The target framework for the project.",
-      "datatype": "choice",
-      "choices": [
-        {
-          "choice": "net9.0",
-          "description": "Target net9.0",
-          "displayName": ".NET 9.0"
-        }
-      ],
-      "defaultValue": "net9.0",
-      "replaces": "netX.0",
-      "displayName": "framework"
-    }
-  }
-}
