--- src/mono/wasm/templates/templates/wasi-native/Properties/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-[assembly:System.Runtime.Versioning.SupportedOSPlatform("wasi")]
