--- src/mono/wasm/templates/templates/wasi-native/runtimeconfig.template.json
@@ -1,10 +1,0 @@
-{
-    "wasmHostProperties": {
-        "perHostConfig": [
-            {
-                "name": "wasmtime",
-                "Host": "wasmtime"
-            }
-        ]
-    }
-}
