--- src/mono/wasm/templates/templates/wasi-native/wasi-native.0.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>netX.0</TargetFramework>
-    <RuntimeIdentifier>wasi-wasm</RuntimeIdentifier>
-    <OutputType>Exe</OutputType>
-    <WasmBuildNative>true</WasmBuildNative>
-    <WasmNativeStrip>false</WasmNativeStrip>
-    <IsBrowserWasmProject>false</IsBrowserWasmProject>
-    <WasmSingleFileBundle>true</WasmSingleFileBundle>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <ItemGroup>
-    <NativeFileReference Include="local.c" />
-  </ItemGroup>
-</Project>
